+++ _wotlk/Interface/Addons/TradeSkillMaster/Core/UI/Elements/CraftingQueueList.lua
@@ -17,7 +17,6 @@ local RecipeString = TSM.Include("Util.RecipeString")
 local Theme = TSM.Include("Util.Theme")
 local TextureAtlas = TSM.Include("Util.TextureAtlas")
 local ScriptWrapper = TSM.Include("Util.ScriptWrapper")
-local ItemInfo = TSM.Include("Service.ItemInfo")
 local Profession = TSM.Include("Service.Profession")
 local BagTracking = TSM.Include("Service.BagTracking")
 local UIUtils = TSM.Include("UI.UIUtils")
@@ -273,7 +272,7 @@ function private.GetItemText(self, data)
 		local craftString = CraftString.FromRecipeString(recipeString)
 		local itemString = TSM.Crafting.GetItemString(craftString)
 		local spellId = CraftString.GetSpellId(craftString)
-		return itemString and UIUtils.GetColoredItemName(itemString) or GetSpellInfo(spellId) or "?"
+		return itemString and UIUtils.GetColoredCraftedItemName(itemString) or GetSpellInfo(spellId) or "?"
 	end
 end
 
@@ -294,7 +293,7 @@ function private.GetItemTooltip(self, data)
 	local numQueued = data:GetField("num")
 	local itemString = TSM.Crafting.GetItemString(craftString)
 	local spellId = CraftString.GetSpellId(craftString)
-	local name = itemString and UIUtils.GetColoredItemName(itemString) or GetSpellInfo(spellId) or "?"
+	local name = itemString and UIUtils.GetColoredCraftedItemName(itemString) or GetSpellInfo(spellId) or "?"
 	local tooltipLines = TempTable.Acquire()
 	tinsert(tooltipLines, name.." (x"..numQueued..")")
 	local numResult = TSM.Crafting.GetNumResult(craftString)
@@ -306,17 +305,18 @@ function private.GetItemTooltip(self, data)
 		local numHave = BagTracking.GetCraftingMatQuantity(matItemString)
 		local numNeed = quantity * numQueued
 		local color = Theme.GetColor(numHave >= numNeed and "FEEDBACK_GREEN" or "FEEDBACK_RED")
-		tinsert(tooltipLines, color:ColorText(numHave.."/"..numNeed).." - "..(ItemInfo.GetName(matItemString) or "?"))
+		tinsert(tooltipLines, color:ColorText(numHave.."/"..numNeed).." - "..(UIUtils.GetColoredCraftedItemName(matItemString) or "?"))
 	end
 	for _, _, itemId in RecipeString.OptionalMatIterator(recipeString) do
 		local matItemString = "i:"..itemId
 		local numHave = BagTracking.GetCraftingMatQuantity(matItemString)
 		local numNeed = TSM.Crafting.GetOptionalMatQuantity(craftString, itemId)* numQueued
 		local color = Theme.GetColor(numHave >= numNeed and "FEEDBACK_GREEN" or "FEEDBACK_RED")
-		tinsert(tooltipLines, color:ColorText(numHave.."/"..numNeed).." - "..(ItemInfo.GetName(matItemString) or "?"))
+		tinsert(tooltipLines, color:ColorText(numHave.."/"..numNeed).." - "..(UIUtils.GetColoredCraftedItemName(matItemString) or "?"))
 	end
-	if Profession.GetRemainingCooldown(craftString) then
-		tinsert(tooltipLines, Theme.GetColor("FEEDBACK_RED"):ColorText(L["On Cooldown"]))
+	local cooldown = Profession.GetRemainingCooldown(craftString)
+	if cooldown then
+		tinsert(tooltipLines, Theme.GetColor("FEEDBACK_RED"):ColorText(format(L["On Cooldown for %s"], SecondsToTime(cooldown))))
 	end
 	return strjoin("\n", TempTable.UnpackAndRelease(tooltipLines)), true, true
 end