+++ _wotlk/Interface/Addons/TradeSkillMaster/ChangeLog.md
@@ -1,5 +1,7 @@
-## v4.12.17 Changes
+## v4.12.18 Changes
 
-* [All] Various Crafting bug fixes
+* [All] Fixed varuious bugs / issues with recipes on cooldown
+* [All] Improved performance when opening profession with TSM
+* [Retail] Show quality in crafting queue tooltip
 
 [Known Issues](https://support.tradeskillmaster.com/en_US/known_issues)